Extract auth logic into a separated singleton
