+++ lib/supabase_flutter.dart
@@ -4,5 +4,6 @@
 library supabase_flutter;
 
 export 'src/supabase.dart';
+export 'src/supabase_auth.dart';
 export 'src/supabase_auth_required_state.dart';
 export 'src/supabase_auth_state.dart';
